Lightly refactor CommandResolutionHelper
